Rename module to match package name
