+++ django_tasks/apps.py
@@ -2,7 +2,7 @@ from django.apps import AppConfig
 
 
 class TasksAppConfig(AppConfig):
-    name = "django_core_tasks"
+    name = "django_tasks"
 
     def ready(self) -> None:
         from . import signal_handlers  # noqa
