+++ tests/settings.py
@@ -9,7 +9,7 @@ BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 ALLOWED_HOSTS = ["*"]
 
 INSTALLED_APPS = [
-    "django_core_tasks",
+    "django_tasks",
     "tests",
 ]
